playwright/requirements.txt
pytest==7.4.4
playwright==1.41.0
pytest-playwright==0.4.4
pytest-asyncio==0.21.1
python-dotenv==1.0.0
requests==2.31.0